# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/user/secondaryEducation": {
            "get": {
                "tags": [
                    "Secondary Education"
                ],
                "summary": "Get all secondary education records",
                "operationId": "813d0f56aa5e620135fcb985109ce95b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SecondaryEducation"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Secondary Education"
                ],
                "summary": "Store new secondary education record",
                "operationId": "3e879ce0448713ffc1760bf9189a2b40",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "school",
                                    "kcseYear",
                                    "grade"
                                ],
                                "properties": {
                                    "school": {
                                        "type": "string"
                                    },
                                    "kcseYear": {
                                        "type": "integer"
                                    },
                                    "grade": {
                                        "type": "string"
                                    },
                                    "kcseCertificate": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Secondary education record created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SecondaryEducation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/user/secondaryEducation/{id}": {
            "get": {
                "tags": [
                    "Secondary Education"
                ],
                "summary": "Get specific secondary education record",
                "operationId": "2f7bb8e02047ee52c9f6da2f4148852f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SecondaryEducation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Secondary Education"
                ],
                "summary": "Update specific secondary education record",
                "operationId": "dcde014681a972f8ac34e2b93e8d599d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "school",
                                    "kcseYear",
                                    "grade"
                                ],
                                "properties": {
                                    "school": {
                                        "type": "string"
                                    },
                                    "kcseYear": {
                                        "type": "integer"
                                    },
                                    "grade": {
                                        "type": "string"
                                    },
                                    "kcseCertificate": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Secondary education record updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SecondaryEducation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Secondary Education"
                ],
                "summary": "Delete specific secondary education record",
                "operationId": "1af47ea0fe6190b249110a5349db9d55",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Secondary education record deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found"
                    }
                }
            }
        },
...
    },
...
}